11ty config is similier to follow - addPlugin for transforms --- this is where esbuild, sass, and html min happen - addPlugin for filters --- adds human readable dates - TODO addShortcodes
