+++ package.json
@@ -28,6 +28,7 @@
     "esbuild": "^0.14.38",
     "esbuild-sass-plugin": "^2.2.6",
     "html-minifier": "^4.0.0",
+    "moment-timezone": "^0.5.34",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.6",
     "postcss-preset-env": "^7.5.0",
