--- src/__flightdeck/afterBuild.js
@@ -1,9 +1,0 @@
-/* cSpell:disable */
-const esbuildBundle = requre("./afterBuild/esbuild.js")
-
-module.exports =  (config) => {
-
-  if (process.env.ELEVENTY_ENV === "production") {
-    config.addTransform("htmlmin", minifyHtml);
-  }
-};
